Move familyaccessioningstackedtable
